Clean up file structure.
