+++ footer.php
@@ -1,4 +1,7 @@
-<?php if (get_theme_mod('footer_sticky') === 'enabled') : ?>
+<?php
+use \Tofino\ThemeOptions\Notifications as n;
+
+if (get_theme_mod('footer_sticky') === 'enabled') : ?>
   </div>
 <?php endif; ?>
 
@@ -16,7 +19,7 @@
 
 <?php wp_footer(); ?>
 
-<?php \Tofino\ThemeOptions\notification('bottom'); ?>
+<?php n\notification('bottom'); ?>
 
 </body>
 </html>
