+++ page.php
@@ -12,9 +12,11 @@
  * @since 1.0.0
  */
 
+use \Tofino\Helpers as h;
+
 get_header();
 
-$template = Tofino\Helpers\get_page_name();
+$template = h\get_page_name();
 
 if (locate_template('templates/content-page-' . $template . '.php') != '') {
   get_template_part('templates/content-page', $template); // e.g. templates/content-page-members.php
@@ -22,7 +24,7 @@ if (locate_template('templates/content-page-' . $template . '.php') != '') {
   if (function_exists('icl_object_id')) { //WPML installed
     $original_page_id = apply_filters('wpml_object_id', get_the_ID(), 'page', false, 'en'); //Assumes english is the primary language.
     if ($original_page_id) {
-      $template = Tofino\Helpers\get_page_name($original_page_id);
+      $template = h\get_page_name($original_page_id);
     }
   }
   get_template_part('templates/content-page', $template);
