+++ src/lib/init.php
@@ -60,30 +60,6 @@ add_action('after_setup_theme', __NAMESPACE__ . '\\check_page_display');
 
 
 /**
- * Show maintenance mode message in admin area
- *
- * Check the maintenance_mode_enabled Theme Option. If enabled display a notice in
- * the admin area at the top of every screen. Also show a popup window to the user
- * and set a cookie.
- *
- * @since 1.0.0
- * @return void
- */
-function show_maintenance_message() {
-  if (get_theme_mod('maintenance_mode') === true) {?>
-    <div class="error notice">
-      <p><strong><?php echo __('Maintenance Mode', 'tofino') . '</strong> - ' . get_theme_mod('maintenance_mode_text', __('This site is currently in maintenance mode. Any changes you make may be overwritten or removed.', 'tofino')); ?></p>
-    </div><?php
-
-    if (!isset($_COOKIE['tofino_maintenance_alert_dismissed'])) {
-      echo '<div class="maintenance-mode-alert"><h1>' . __('Maintenance Mode', 'tofino') . '</h1><p>' . get_theme_mod('maintenance_mode_text', __('This site is currently in maintenance mode. Any changes you make may be overwritten or removed.', 'tofino')) . '</p><button>' . __('I understand', 'tofino') . '</button></div>';
-    }
-  }
-}
-add_action('admin_notices', __NAMESPACE__ . '\\show_maintenance_message');
-
-
-/**
  * Set max content width GLOBAL
  *
  * @since 1.0.0
